+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/implementation/directconnectivity/StoreReaderTest.java
@@ -37,7 +37,9 @@ import reactor.core.publisher.DirectProcessor;
 import reactor.core.publisher.Mono;
 
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.TimeUnit;
@@ -636,13 +638,15 @@ public class StoreReaderTest {
                 .withGlobalCommittedLsn(bigLsn)
                 .build();
 
+        Map<String, List<String>> replicaStatusList = new HashMap<>();
+        replicaStatusList.put(Uri.ATTEMPTING, Arrays.asList(primaryURI.getHealthStatusDiagnosticString()));
         StoreResult result = storeReader.createStoreResult(
                 storeResponse,
                 null,
                 false,
                 false,
                 null,
-                Arrays.asList(primaryURI.getHealthStatusDiagnosticString()));
+                 replicaStatusList);
         assertThat(result.globalCommittedLSN).isEqualTo(bigLsn);
         assertThat(result.lsn).isEqualTo(bigLsn);
     }